change int to ssize_t constant

+++ 0x15-file_io/main.h
@@ -1,6 +1,12 @@
 #ifndef _MAIN_H_
 #define _MAIN_H_
 
-int read_textfile(const char *filename, int letters);
+#include <stdio.h>
+#include <stdlib.h>
+#include <fcntl.h>
+#include <unistd.h>
+
+
+ssize_t read_textfile(const char *filename, size_t letters);
 
 #endif
